all components now import from multi-lingual feature
