+++ ADP.WebComponents/adp-web-components/src/components/part-lookup/dead-stock-lookup.tsx
@@ -1,16 +1,15 @@
 import { InferType } from 'yup';
 import { Component, Element, Host, Method, Prop, State, Watch, h } from '@stencil/core';
 
-import { LanguageKeys } from '~types/locale';
 import { AppStates, MockJson } from '~types/components';
 import { PartLookupDTO } from '~types/generated/part/part-lookup-dto';
 
 import cn from '~lib/cn';
-import { ErrorKeys, getLocaleLanguage, getSharedLocal, SharedLocales, sharedLocalesSchema } from '~lib/get-local-language';
 
 import { getPartInformation, PartInformationInterface } from '~api/partInformation';
 
 import deadStockSchema from '~locales/partLookup/deadStock/type';
+import { ErrorKeys, getLocaleLanguage, getSharedLocal, LanguageKeys, SharedLocales, sharedLocalesSchema } from '~features/multi-lingual';
 
 let mockData: MockJson<PartLookupDTO> = {};
