+++ ADP.WebComponents/adp-web-components/src/components/vehicle-lookup/vehicle-lookup.tsx
@@ -1,9 +1,6 @@
 import { Component, Element, Host, Method, Prop, State, Watch, h } from '@stencil/core';
 
 import validateVin from '~lib/validate-vin';
-import { ErrorKeys, getLocaleLanguage, getSharedLocal, SharedLocales, sharedLocalesSchema } from '~lib/get-local-language';
-
-import { LanguageKeys } from '~types/locale';
 import { DotNetObjectReference } from '~types/components';
 import { VehicleLookupDTO } from '~types/generated/vehicle-lookup/vehicle-lookup-dto';
 
@@ -17,6 +14,8 @@ import { VehicleServiceHistory } from './vehicle-service-history';
 import { VehicleWarrantyDetails } from './vehicle-warranty-details';
 
 import { VehicleInfoLayout } from '../components/vehicle-info-layout';
+
+import { ErrorKeys, getLocaleLanguage, getSharedLocal, LanguageKeys, SharedLocales, sharedLocalesSchema } from '~features/multi-lingual';
 
 const componentTags = {
   vehicleAccessories: 'vehicle-accessories',
