+++ ADP.WebComponents/adp-web-components/src/features/multi-lingual/index.ts
@@ -1,0 +1,3 @@
+export * from './types';
+export * from './interface';
+export * from './get-local-language';
