+++ ADP.WebComponents/adp-web-components/src/features/multi-lingual/interface.ts
@@ -1,0 +1,11 @@
+import { AnyObjectSchema } from 'yup';
+
+import { ComponentLocale, LanguageKeys } from './types';
+
+export interface MultiLingual {
+  language: LanguageKeys;
+  locale: ComponentLocale<AnyObjectSchema>;
+
+  componentWillLoad: () => Promise<void>;
+  changeLanguage: (newLanguage: LanguageKeys) => Promise<void>;
+}
