--- ADP.WebComponents/adp-web-components/src/global/lib/component-locale.ts
@@ -1,8 +1,0 @@
-import { AnyObjectSchema, InferType } from 'yup';
-import { SharedLocales } from './get-local-language';
-
-type ComponentLocale<T extends AnyObjectSchema> = {
-  sharedLocales?: SharedLocales;
-} & InferType<T>;
-
-export default ComponentLocale;
