--- ADP.WebComponents/adp-web-components/src/global/types/locale.ts
@@ -1,13 +1,0 @@
-export const ARABIC_JSON_FILE = 'ar.json';
-export const ENGLISH_JSON_FILE = 'en.json';
-export const KURDISH_JSON_FILE = 'ku.json';
-export const RUSSIAN_JSON_FILE = 'ru.json';
-
-export type LanguageKeys = 'en' | 'ar' | 'ku' | 'ru';
-
-export const languageMapper = {
-  ar: ARABIC_JSON_FILE,
-  en: ENGLISH_JSON_FILE,
-  ku: KURDISH_JSON_FILE,
-  ru: RUSSIAN_JSON_FILE,
-};
